+++ packages/bottom-tabs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-alpha.50](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@8.0.0-alpha.49...@react-navigation/bottom-tabs@8.0.0-alpha.50) (2026-08-26)
+
+**Note:** Version bump only for package @react-navigation/bottom-tabs
+
 # [8.0.0-alpha.49](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@8.0.0-alpha.48...@react-navigation/bottom-tabs@8.0.0-alpha.49) (2026-08-19)
 
 ### Bug Fixes