+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-alpha.51](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@8.0.0-alpha.50...@react-navigation/drawer@8.0.0-alpha.51) (2026-08-26)
+
+**Note:** Version bump only for package @react-navigation/drawer
+
 # [8.0.0-alpha.50](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@8.0.0-alpha.49...@react-navigation/drawer@8.0.0-alpha.50) (2026-08-19)
 
 ### Bug Fixes