+++ packages/elements/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-alpha.48](https://github.com/react-navigation/react-navigation/compare/@react-navigation/elements@3.0.0-alpha.47...@react-navigation/elements@3.0.0-alpha.48) (2026-08-26)
+
+**Note:** Version bump only for package @react-navigation/elements
+
 # [3.0.0-alpha.47](https://github.com/react-navigation/react-navigation/compare/@react-navigation/elements@3.0.0-alpha.46...@react-navigation/elements@3.0.0-alpha.47) (2026-08-19)
 
 **Note:** Version bump only for package @react-navigation/elements