+++ packages/stack/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-alpha.53](https://github.com/react-navigation/react-navigation/compare/@react-navigation/stack@8.0.0-alpha.52...@react-navigation/stack@8.0.0-alpha.53) (2026-08-26)
+
+**Note:** Version bump only for package @react-navigation/stack
+
 # [8.0.0-alpha.52](https://github.com/react-navigation/react-navigation/compare/@react-navigation/stack@8.0.0-alpha.51...@react-navigation/stack@8.0.0-alpha.52) (2026-08-19)
 
 ### Bug Fixes